refactor: split spending breakdown components
