+++ frontend/src/accounts/detail/AccountDetailPage.tsx
@@ -10,7 +10,8 @@ import BackLink from '@/accounts/detail/components/BackLink'
 import BalanceChartCard from '@/accounts/detail/components/balance-chart/BalanceChartCard'
 import EditAccountIdentityModal from '@/accounts/detail/components/EditAccountIdentityModal'
 import MonthlyCashFlowCard from '@/accounts/detail/components/monthly-cash-flow/MonthlyCashFlowCard'
-import { TopCategoriesBySpendingCard, TopMerchantsBySpendingCard } from '@/accounts/detail/components/SpendingBreakdownCards'
+import { TopCategoriesBySpendingCard } from '@/accounts/detail/components/spending-breakdown/TopCategoriesBySpendingCard'
+import { TopMerchantsBySpendingCard } from '@/accounts/detail/components/spending-breakdown/TopMerchantsBySpendingCard'
 import { EASE } from '@/accounts/detail/constants/accountDetail'
 import TransactionListSection from '@/transactions/components/TransactionListSection'
